# docs/swagger.yaml
definitions:
...
  errors.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
  models.FeatureTypeEnum:
    enum:
    - static
    - metered
    type: string
    x-enum-varnames:
    - FeatureTypeStatic
    - FeatureTypeMetered
...
  models.HttpResponse-array_models_PlanFeature:
    properties:
      data:
        items:
          $ref: '#/definitions/models.PlanFeature'
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.HttpResponse-models_PlanFeature:
    properties:
      data:
        $ref: '#/definitions/models.PlanFeature'
      message:
        type: string
      status:
        type: integer
    type: object
  models.HttpResponse-models_PlanFeatureQuota:
    properties:
      data:
        $ref: '#/definitions/models.PlanFeatureQuota'
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.MeteredActionAtLimit:
    enum:
    - none
    - block
    - throttle
    type: string
    x-enum-varnames:
    - MeteredActionAtLimitNone
    - MeteredActionAtLimitBlock
    - MeteredActionAtLimitThrottle
  models.MeteredResetPeriod:
    enum:
    - day
    - week
    - month
    - year
    - custom
    - rolling
    - never
    type: string
    x-enum-varnames:
    - MeteredResetPeriodDay
    - MeteredResetPeriodWeek
    - MeteredResetPeriodMonth
    - MeteredResetPeriodYear
    - MeteredResetPeriodCustom
    - MeteredResetPeriodRolling
    - MeteredResetPeriodNever
...
  models.PlanFeature:
    properties:
      config: {}
      created_at:
        type: string
      created_by:
        type: string
      feature_id:
        type: string
      feature_name:
        type: string
      feature_slug:
        type: string
      feature_type:
        $ref: '#/definitions/models.FeatureTypeEnum'
      id:
        type: string
      plan_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
  models.PlanFeatureQuota:
    properties:
      action_at_limit:
        $ref: '#/definitions/models.MeteredActionAtLimit'
      created_at:
        type: string
      created_by:
        type: string
      custom_period_minutes:
        type: integer
      id:
        type: string
      limit_value:
        type: integer
      plan_feature_id:
        type: string
      reset_period:
        $ref: '#/definitions/models.MeteredResetPeriod'
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  planfeatures.createPlanFeatureRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      created_by:
        type: string
      feature_id:
        type: string
    required:
    - created_by
    - feature_id
    type: object
  planfeatures.updatePlanFeatureRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      updated_by:
        type: string
    required:
    - updated_by
    type: object
...
  quotas.createQuotaRequest:
    properties:
      action_at_limit:
        enum:
        - none
        - block
        - throttle
        type: string
      custom_period_minutes:
        type: integer
      limit_value:
        type: integer
      reset_period:
        enum:
        - day
        - week
        - month
        - year
        - custom
        - rolling
        - never
        type: string
    required:
    - action_at_limit
    - limit_value
    - reset_period
    type: object
  quotas.updatePlanFeatureQuotaRequest:
    properties:
      action_at_limit:
        enum:
        - none
        - block
        - throttle
        type: string
      custom_period_minutes:
        type: integer
      limit_value:
        type: integer
      reset_period:
        enum:
        - day
        - week
        - month
        - year
        - custom
        - rolling
        - never
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/plans/{planID}/features:
    get:
      consumes:
      - application/json
      description: Retrieves all features associated with a plan
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Filter by feature type
        in: query
        name: feature_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of plan features
          schema:
            $ref: '#/definitions/models.HttpResponse-array_models_PlanFeature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: List plan features
      tags:
      - Plan Features
    post:
      consumes:
      - application/json
      description: Adds a new feature to a plan
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Plan feature details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/planfeatures.createPlanFeatureRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Plan feature created successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_PlanFeature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "409":
          description: Plan feature already exists
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create plan feature
      tags:
      - Plan Features
  /v1/plans/{planID}/features/{featureID}:
    delete:
      consumes:
      - application/json
      description: Removes a feature from a plan
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Feature ID
        format: uuid
        in: path
        name: featureID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Plan feature deleted successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Plan feature not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Delete plan feature
      tags:
      - Plan Features
    put:
      consumes:
      - application/json
      description: Updates configuration of a feature associated with a plan
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Feature ID
        format: uuid
        in: path
        name: featureID
        required: true
        type: string
      - description: Updated plan feature details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/planfeatures.updatePlanFeatureRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Plan feature updated successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_PlanFeature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Plan feature not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      tags:
      - Plan Features
  /v1/plans/{planID}/features/{featureID}/quotas:
    delete:
      consumes:
      - application/json
      description: Removes the quota configuration from a plan feature
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Feature ID
        format: uuid
        in: path
        name: featureID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Quota deleted successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Quota not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Delete plan feature quota
      tags:
      - Plan Feature Quotas
    get:
      consumes:
      - application/json
      description: Retrieves quota configuration for a specific plan feature
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Feature ID
        format: uuid
        in: path
        name: featureID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Quota details retrieved successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_PlanFeatureQuota'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Quota not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Get plan feature quota details
      tags:
      - Plan Feature Quotas
    post:
      consumes:
      - application/json
      description: Adds a quota configuration to a plan feature
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Feature ID
        format: uuid
        in: path
        name: featureID
        required: true
        type: string
      - description: Quota configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/quotas.createQuotaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Quota created successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_PlanFeatureQuota'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "409":
          description: Quota already exists
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Create plan feature quota
      tags:
      - Plan Feature Quotas
    put:
      consumes:
      - application/json
      description: Updates the quota configuration for a plan feature
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Feature ID
        format: uuid
        in: path
        name: featureID
        required: true
        type: string
      - description: Updated quota configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/quotas.updatePlanFeatureQuotaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Quota updated successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_PlanFeatureQuota'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Quota not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Update plan feature quota
      tags:
      - Plan Feature Quotas
...
swagger: "2.0"